move models to a separate directory and change the import commands in projeto.py
